requirements.txt
Pillow>=9.0.0
reportlab>=3.6.0
pyinstaller>=5.0.0
